requirements.txt
streamlit==1.32.0
groq==1.1.2
SpeechRecognition==3.10.1
python-dotenv==1.0.1
audio-recorder-streamlit==0.0.8
pydub==0.25.1